codionp/requirement.txt
pandas
numpy
torch
matplotlib
Pillow
tqdm
resnet